reorganize helper functions into their own file
